add projects/tests/test_status.py projects/views_status.py
